test: add supertest for testing express app
